Chore - revert startTransformByQ refactor
